backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0  # Compatibility with passlib
python-multipart==0.0.6
authlib==1.3.0
feedparser==6.0.11
aiohttp==3.9.1
redis==5.0.1
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3
openai==1.10.0
anthropic==0.18.1
python-dotenv==1.0.0
httpx==0.26.0

# OCR and scraping
pytesseract==0.3.10
Pillow==10.2.0
beautifulsoup4==4.12.3
pypdf==4.0.1
playwright==1.41.2

# Text processing
python-Levenshtein==0.25.0
fuzzywuzzy==0.18.0
textstat==0.7.3
spacy==3.7.4
vaderSentiment==3.3.2
